requirements.txt
asgiref==3.8.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorclass==2.2.2
contourpy==1.2.1
crispy-bootstrap5==2024.2
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.3
django-crispy-forms==2.1
django-crontab==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.50.0
html5lib==1.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip-upgrader==1.4.15
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.9.0
pytz==2023.3.post1
reportlab==4.1.0
requests==2.31.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
terminaltables==3.1.10
tinycss2==1.2.1
tzdata==2024.1
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
